% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specstats.R
\name{spec_stats}
\alias{spec_stats}
\title{Spec stats}
\usage{
spec_stats()
}
\arguments{
\item{none}{none}
}
\description{
Special statistics of Fide players
}
\examples{
spec_stats()
}
\keyword{Fide}
\keyword{list}
\keyword{players}
\keyword{special}
\keyword{statistics}

